feat: add notes to unpublished section on note creation
